DonorQuestApi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.2.1
SQLAlchemy==2.0.35
psycopg[binary]==3.2.3
passlib>=1.7.4
bcrypt<4
python-jose==3.3.0
python-multipart==0.0.12
firebase-admin==6.5.0
requests==2.32.3
python-dotenv==1.0.1
slowapi==0.1.9
geographiclib==2.0
pandas==2.2.3
email-validator==2.0.0.post2
pip==23.2.1
